Moved AutomatedAssessment from structure to add-node.
Changed AutomatedAssessment behavior so that it opens the node authoring view when it's added to the unit.
#54
