+++ src/assets/wise5/authoringTool/addNode/add-your-own-node/add-your-own-node.component.html
@@ -5,11 +5,10 @@
   </mat-form-field>
 </div>
 <hr>
-<div fxLayout="row" fxLayoutAlign="end center">
-  <button mat-button class="mat-primary"
-      (click)="cancel()"
-      aria-label="cancel"
-      i18n>Cancel</button>
+<div fxLayout="row">
+  <button mat-button class="mat-primary" (click)="back()" aria-label="back" i18n>Back</button>
+  <span fxFlex></span>
+  <button mat-button class="mat-primary" (click)="cancel()" aria-label="cancel" i18n>Cancel</button>
   <button mat-button class="mat-raised-button mat-primary"
       (click)="chooseLocation()"
       aria-label="next"
