+++ src/assets/wise5/authoringTool/addNode/automatedAssessment/choose-item.html
@@ -25,7 +25,7 @@
 <hr>
 <div layout="row">
   <md-button class="createButton md-primary"
-             ng-click="$ctrl.goToChooseStructure()"
+             ng-click="$ctrl.back()"
              aria-label="{{ ::'back' | translate }}"
              translate="back"></md-button>
   <span flex></span>
